Add grunt-compass to comple sass files
Grunt-compass was added to allow JS and CSS files to be concatenated
and minified using a single terminal with "grunt watch". This allows
to separate javascript and css files into different files when needed
but only load the optimized version on runtime.

Because of this files have a different structure:

- Only javascript files related to the project are stored in the "js"
folder, third party libraries and plugins are stored in the "libs"
folder.
- Sass files (.scss) are stored in the "sass" folder, individual styles
files are stored in the "partials" folder and are prefixed with a
single underscore.
- Grunt's config files are stored in the "src" folder.